finish logic for window hook
